removed settings
